Community conversation: New chat panel view (#4178) (#4336)
* Added topics or community id to the multiplayer messages

It requires the latest version of our livekit library.
For now, it is sending custom-topic in the private messages.

* Using the new livekit package version from git

Added a new prefab to show the title of a community conversation.
Added a new prefab to show the conversation icon in the chat conversation toolbar.
Now adding a channel has been split into 3 to allow different type of data to be provided depending on the type of conversation.
IProfileThumbnailCache renamed to IThumbnailCache, as an attempt to make it more generic. IThumbnailCache is used for getting the image of the community asynchronously.
Now the multiplayer bus handles community conversation messages too.
Now community messages send a topic along with the messages. Messages are sent to a special user created server-side which resends them to members of the community specified in the topic.
Now a new conversation is added to the chat when the user logs in for each community he belongs to.
Multiple users can communicate using a community conversation.

* The name of the routing user now depends on the environment

* Added context menu for the title bar

* View community, Add community conversation and remove community conversation