Feat: sdk scene video and streams fade in/out (#4418)
# Pull Request Description
This PR implements audio fade in/out for audio from LiveKit streams and MediaPlayer video/audio.

- extended logic to fade instead to set volume directly to the final value

## Test Instructions
Test on scenes with Video streams (like Game Night or `/goto 1,93`).
If possible, try to test it with the Live Streaming.

### Test Steps 1
GIVEN player is on the SDK scene with Video stream or Live Stream
WHEN player step out of the scene by foot
THEN audio should slowly fade out (and not just be off/mute in one hit)

### Test Steps 2
GIVEN player is near to the SDK scene with Video stream or Live Stream
WHEN player step in to the scene by foot
THEN audio should slowly fade in (and not just be on on full volume)