Now the Member list is reused in both Nearby and Communities
It shows the actual members in the list in each community.
It updates the number of members dynamically.
The member list data structure has been renamed to ChatUserData, so it does not matter the origin of the data (communities endpoints or profile repository), the same method and data layout is sent to the member list and to the input box for comparing the cached profiles with the members.