Feat: loading screen - avatar head look at mouse cursor (IK) (#4403)
# Pull Request Description

Implemented head look to mouse cursor IK animation for Avatar Previews (includes all Previews - Auth Screen, Backpack and Passport). 

## What does this PR change?
- add system to rotate head towards cursor for Avatar Previews
- introduces settings to tweak head angle
- blocks head rotation when avatar performs emote

## Test Instructions
original task was only for Auth Preview, but it works also in Backpack and Passport, so please test it in all 3 modes. Try change wearables, check if head rotates when emoting (it shouldn't). 

Specifically, check if you can get to such screwed state (it was discovered in early stage of development and could be fixed). This is probably specific to backpack, so try to change gender and wearables. If you find this issue, try again from the clean start to repro it only on Preview and Passport to isolate issue to Backpack (maybe we will disable it for backpack)

https://github.com/user-attachments/assets/6489589b-f224-462f-80a0-27a886182abb

### Test Steps 1 (happy path - look at cursor)
1. GIVEN you are in the Auth screen, Backpack or Passport UI
2. WHEN you move mouse cursor
3. **THEN** avatar should look towards your mouse position

### Test Steps 2 (edge case 1 - not played during emoting)
1. GIVEN you are in the Passport UI
2. AND avatar is performing waving emot
3. WHEN you move mouse cursor
4. **THEN** avatar should NOT look towards your mouse position while performing emote 
5. **AND** only starts to look this direction after emote is finished

### Test Steps 3 (edge case 2 - try rotate avatar preview)
1. GIVEN you are in the Auth screen, Backpack or Passport UI
2. AND you rotated avatar so it stays by its back or shoulders to your screen
3. WHEN you move mouse cursor
3. **THEN** avatar should try to look towards your mouse position, but not "break his neck"