First implementation of the sprite cache, a refactor of IThumbnailCache
IThumbnailCache renamed to ISpriteCache and moved to UI.
Used in the ProfilePictureView and the ImageController.
The ImageController can use either a web request controller or a sprite cache.
It only works with URLs, not IDs.
The cache of the Community Browser is shared with Community Card Controller when opening a card. Otherwise the Community Card Controller creates its own.
An ObjectProxy is used so the reference to the proxy is stored when the views are instantiated, and its content is set afterwards during the OnViewShow.
Added a Clear method to SpriteCache which empties all the internal dictionaries, called when the Community Browser is closed, or when the Community Card is closed if it uses its own cache.