Fix: url check for avatar profiles (#4731)
# Pull Request Description
Relates to #4661

Cases for the profile picture fetching could be related to the loading going to irrecoverable flow due to 404 return from server (stopping attempts). To resolve it, I marked Promise to be as AvatarTexture since it skips 404 and has delay between retries. 

 Additionally 
 - added URL check when creating profile fetching promise. 
 - added more logs info for better awareness from where `undefined` is coming

## Test Instructions
Make a smoke test regarding profile pictures. Note if there are any delays or problems with it. In general, profile pictures should load as usual. Clear cache before testing.