Chat improvements: Offline Messages and Players (#4719)
Now the ChatController stores which users in nearby and in community channels are online.
Every time a channel is selected in the chat panel, the list of online users corresponding to that channel is sent to the view so it passes it to the messages viewer, which uses that list to redraw the current visible message bubbles applying a greyout to those what belongs to users that are not present in the list (i.e. offline).
A GreyOut method has been added to the classes of the UI hierarchy so they can apply the greyout effect.
Now the connection status indicator of the toolbar icons only appear when the user is online.
The profile picture of the private conversation icons in the toolbar are greyed out too when offline.
ChatEntry prefabs modified to provide access to some UI elements that are affected by the greyout effect.