Private conversations are now properly greyed out
The online private conversations are stored in a special hashset in the participants per channel dictionary.
The profile picture in the toolbar and in the titlebar are updated accordingly.
Messages in private conversations are greyed out.
Known limitation: When the user A enters DCL while it has a private conversation open with user B, and B is already connected, the icon in the toolbar looks bad.