Fix: Avatar Rotation | Avatars not facing correct direction during streaming events (#4789)
Pull Request Description
Fix #3935

Rotation wasn't taken into account when setting initial position of avatars, or teleporting it.
I added rotation assignment in places where we assign position.

Test Steps
GIVEN there are some people/avatars already in the game (in some scene or world)
AND they you are not moving they avatars (for example, watching the stream)
WHEN you logged in to this place
THEN you see avatars rotated correctly (in the direction they are actually looking and not blank North)