Chat improvements: Offline messages and players (#4792)
* Chat improvements: Offline Messages and Players (#4719)

Now the ChatController stores which users in nearby and in community channels are online.
Every time a channel is selected in the chat panel, the list of online users corresponding to that channel is sent to the view so it passes it to the messages viewer, which uses that list to redraw the current visible message bubbles applying a greyout to those what belongs to users that are not present in the list (i.e. offline).
A GreyOut method has been added to the classes of the UI hierarchy so they can apply the greyout effect.
Now the connection status indicator of the toolbar icons only appear when the user is online.
The profile picture of the private conversation icons in the toolbar are greyed out too when offline.
ChatEntry prefabs modified to provide access to some UI elements that are affected by the greyout effect.

* Image coloring replaced with canvas groups' alpha for message and username

* Fixes and improvements

Enabled raycast block in canvas groups.
Fixed: The online users were not being updated when the user left Nearby.
Messages are now manually updated.

* Private conversations are now properly greyed out

The online private conversations are stored in a special hashset in the participants per channel dictionary.
The profile picture in the toolbar and in the titlebar are updated accordingly.
Messages in private conversations are greyed out.

* Code moved to a separate class UserConnectivityInfoProvider

* Fixed: The index being used in the ApplyGreyOut method was wrong

It didn't take into account the separator.

* Small code reversions

* GreyOut methods now only receive the float value

* Fixed: The profile pictures in the toolbar were not being properly initialized

* Documentation added + minor changes

* Fixed: It was failing when initializing the community icons of the toolbar