Chat improvements: Notification per Conversation (#4870)
* Replaced previous implementation of context menu and added submenu for Notification ping options

Notification ping settings are stored in player prefs, per channel. The default value is the one stored per user.

* Now notification sounds are played according to each channel's settings

Fixed: Chat sounds were not being played at all due to a bug, a missing !.
The ChatContextMenuConfiguration script has been moved to the ContextMenus folder.

* Added some documentation

* Just changed the order of the context menu buttons

* The submenu is now properly aligned

* Update chat bar

* Update title bar transparency color

* Update ChatTitlebarDefault.prefab

* This is a test, REVERT IT

* Fixed some visual styling issues

Added better margins.
The size of the icons is now the specified in the designs.

* Update font style for chat member list button

* Fixed styles of the context menu

* Revert "This is a test, REVERT IT"

This reverts commit 07ca7ee89f95d51f95c122a8a530726cae9fc6ff.

* In theory, fixed the problem with the animation of the mention bubble

When the mention bubble is animated to recover the size of the nametag (when hiding), the size of the bubble was not being changed.
It also includes a fix for an out of bound index when receiving messages (related to timestamps).

---------

Co-authored-by: Romina Marchetti <rma.marchetti@gmail.com>