Minimap: Black screen on first open (#4975)
The problem was that the default tab in the SectionSelectorController was not being set depending on the FF of Communities, however the last shown section variable was. Due to this inconsistency, the Activate method of the NavmapController was not being called when the minimap section was shown the first time the Explore panel is open (the "shownSection == lastShownSection" condition fails, the Activate method is not called there).