Prototype of the idea working in the Chat panel context menu
The submenu is loaded asynchronously when the mouse cursor hovers the corresponding item in the parent context menu. The submenu is rebuilt every time.
The client that opens the menu is in charge of providing a delegate that gives the submenu the appropriate control settings. This will be called from the controller before configuring the submenu.