Now submenu buttons can be hidden asynchronously
When a context menu is loaded, if there is a submenu that may not be available depending on a condition that has to be calculated asynchronously, that button shows a spinner and, once the visibility is calculated, it either shows the button and its submenu or it hides the button from the parent context menu.
Fixed a potential problem in the scroll view menus, they weren't clearing its content when reused, so buttons were accumulating.
Now the submenu is not calling the server, as the data it needs is the same that is gathered when calculating the visibility of the button.