Private communities: Invitations context menu (#5120)
* Implemented asynchronous loading of context menu submenus

The submenu is loaded asynchronously when the mouse cursor hovers the corresponding item in the parent context menu. The submenu is rebuilt every time.
The client that opens the menu is in charge of providing a delegate that gives the submenu the appropriate control settings. This will be called from the controller before configuring the submenu.
Added loading spinner.

* Added Invite to Community async submenu to User profile context menu

* Now the submenu can send invitations

* Now submenu buttons can be hidden asynchronously

When a context menu is loaded, if there is a submenu that may not be available depending on a condition that has to be calculated asynchronously, that button shows a spinner and, once the visibility is calculated, it either shows the button and its submenu or it hides the button from the parent context menu.
Fixed a potential problem in the scroll view menus, they weren't clearing its content when reused, so buttons were accumulating.
Now the submenu is not calling the server, as the data it needs is the same that is gathered when calculating the visibility of the button.

* Now the Invite to community button appears in the passport too

Code refactorized to put everything that has to do with the Invite to Community button in one separate place that can be shared among both user profile context menus.
Added a dependency from Passport to CommunitiesDataProvider and to the CommunitiesContextMenu asmdefs.

* Implemented the call to the correct endpoint

* Code cleaning

* Minor change in a namespace

* Now submenus are always totally inside of the screen

Fixed a problem with the call to the endpoint in the Passport.

* Add communities icon for context menu

* When an invitation is sent, a notification is shown at the top of the screen

* Submenu added to the context menu of the members in the Communities panels

Also updated the icon of the submenu with the one provided in Passport and user profile context menu.

* Fix after merge

* Cleaning and documenting

* Documentation and improvements

* Minor change

* Fixed a small bug

---------

Co-authored-by: Romina Marchetti <rma.marchetti@gmail.com>