Add error notification to New Notifications system (#5174)
The SystemNotification prefab has been refactored to use layouts and adjust to its content and configuration. Now it can hide the close button and its width will adapt to the text.
Using the ServerErrorNotification a card will appear at the top of the screen and not in the sidebar's feed.