Chore: Roads | color instancing regroup (#4946)
# Pull Request Description
Fix #4839 
Fix #4376
and potentially Fix #4377

This PR brings usage of color tinting in the GPU Instancing. To achieve this, we extract same CombinedRenderers from one LODGroup prop to its own separate LODGroup prop with just this renderer but with different colors in the PerInstance buffer. This increases memory consumption but optimizes rendering time. In order to achieve this, refactoring was done to remove coupling of data to MonoBehaviours of original Unity LODGroups.

With this change we save `0.12 ms` on the `DrawOpaqueObjects` in `RenderThread`
<img width="451" height="38" alt="image" src="https://github.com/user-attachments/assets/a101b1b1-2775-4eeb-87e2-cebe9ee99b90" />
<img width="335" height="285" alt="image" src="https://github.com/user-attachments/assets/f4706df0-6f3a-40c5-87cd-e34ff9f34e41" />

However, due to extraction we present some data duplication and memory consumption increased by ~ `2.14 MB`: 
from 54,018 [instances] x 96 [byte] = 4.95 [MB]
to      77,395 [instances] x 96 [byte] = 7.09 [MB]

## What does this PR change?
  - Refactor to separate instancing data from original MonoBehaviours
  - Extracted whitelisted shaders into settings level from prefabs level
  - Extends logic of data collection with extraction of same Renderers with different colors. 
  - Add logic to remove LODGroup prop with no renderers in the list
  - fix color tinting in grass shader (point to the branch in the packages during testing)
  - Set smoothing to 0 in the RoadAssets material for the Pot
  
## Test Instructions
Check #4839 for related testing of the Fix. 

Apart from that, just verify that all road props look same as on prod. Pay attention to colors of the bushes - they should be same and in the same position (i.e. for same bushes)

Check different locations and verify that roads are not missing