Avatars animate synchronously when interacting with a social emote
The armature of the receiver is renamed while the outcome animation is playing.
Added the initiator's wallet address to the data sent through livekit.
Added the SocialEmoteInteractionSystem.
The receiver is teleported to the position of the initiator.
The initiator plays the outcome animation when the receiver reacts to the interaction.
The AvatarBase prefab now stores a reference to the Armature object.
When hovering another avatar with the mouse cursor, it looks different depending on whether it has started an interaction.