Fixed bugs occurring when cancelling a social emote interaction
Protobuf library updated (isStopped and isRepeating added).
Created the StopEmoteIntent component, which is consumed in a query in CharacterEmoteSystem.
Now if the movement of the avatar cancels an emote, is not checked in both clients, instead the field isStopping was added to PlayerEmote and is filled and sent to the other clients, so they cancel the emote when they receive it.
Now when a social emote is cancelled and both avatars were playing the outcome animations, it attaches a StopEmoteIntent to the entity of the other avatar (entity ids are now stored in the interaction state).
Muted the transitions from Emote or EmoteLoop to Movement, in the Animation controller. Movement detection that cancels emotes is checked in conde only.