Outcome animation start pose interpolation working locally
Added a debug utility to draw gizmos from system queries.
Modified the animation controller by adding another transition from Movement to Emote, which does not make any interpolation between animations. Also added bool parameter IsSocialEmoteOutcome that makes the new transition occur when the animation is a social emote outcome.
Now reacting to a social emote makes the avatar run to the position of the initiator. Then it interpolates the position to the future position of the outcome animation's first frame. Then plays the emote. The initiator looks at the reacting avatar.