End outcome animation transition improved
Now both characters stay with the same pose they had when the animation was canceled or complete.
Added a state machine behaviour to the AnimationController of the character in order to know when is the animation state finishing.
Fixed the RotateCharacterSystem.ForceLookAt query, the relative rotation to the platform was not being saved. This was causing that the rotation applied to the character was being overwritten in CharacterPlatformSystem.ResolvePlatformMovement.
The player is not sending movement messages while playing an outcome animation.