Some fixes + Disconnection handling
The MoveBeforePlayingSocialEmoteIntent was being removed in the wrong place so while playing an emote it was not possible to react to other, it cancelled the emote but the avatar didn't walk.
Now StopEmote is called before the same social emote is played.
When the initiator is not available while the avatar is walking towards it, the movement is canceled.
Added the flag HasPlayedEmote to CharacterEmoteIntent so the next query knows if the emote has been played in previous steps or not.
Now it stops interaction when player is disconnected.