Several improvements
Now it shows 2 options when interacting. The other option shows the emote.
Now the outcomes menu appears at the left of the avatar.
Now the cursor is not hidden when showing the outcomes menu in locked mode.
Now the pins are shown atop the avatar to which you can react.