Avatar highlight implemented in a separate system with 2 intents
The highlight opacity decreases every frame unless there is an intent to increase it (when hovering the initiator of a social emote)
Depending on the distance from the receiver to the initiator, the color of the outline varies.
There is a special intent to play a blinking animation, played when a directed social emote arrives. It has priority over the normal highlight.
Includes a Fix for ephemeral notifications not showing when sending an emote while playing another emote.
Moved some intents to the Intents folder.