Fixed problem with avatars rotating suddenly while sitting on a bench
It happened when a user updated the profile. That causes the cached profile data to reset and there is a condition that compared that empty wallet address of that profile with the target of a directed emote operation (which is not happening). Due to both are empty strings it was meeting the condition and making avatars look at an inexistent initiator of a social emote.