Fix: Props and assets intermittently disappear during emotes in crowded scenes (#6467)
Now the CurrentEmoteReference property is set to null when the emote is stopped.
The problem: when stopping the emote, CurrentEmoteReference still has a value despite the instance was returned to the pool, if you play it again CurrentEmoteReference is checked in EmotePlayer.Play() and it reuses it. If other player got that emote instance in the meantime, both are pointing to the same emote which is repositioned every time the loop iterates. With this change, that property is nullified when stopped, and the next time it's played it takes a different instance from the pool.